scripts/requirements.txt
openai
psycopg2-binary
pandas
numpy
scikit-learn
matplotlib
silhouette-score

